+++ BinanceExchange.API/Client/Interfaces/IBinanceClient.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using BinanceExchange.API.Models.Request;
 using BinanceExchange.API.Models.Response;
+using BinanceExchange.API.Models.Response.Abstract;
 
 namespace BinanceExchange.API.Client.Interfaces
 {
@@ -85,7 +86,7 @@ namespace BinanceExchange.API.Client.Interfaces
         /// </summary>
         /// <param name="request">The <see cref="CreateOrderRequest"/> that is used to define the order</param>
         /// <returns></returns>
-        Task<CreateOrderResponse> CreateOrder(CreateOrderRequest request);
+        Task<BaseCreateOrderResponse> CreateOrder(CreateOrderRequest request);
 
         /// <summary>
         /// Queries an order based on the provided request
@@ -149,5 +150,11 @@ namespace BinanceExchange.API.Client.Interfaces
         /// <param name="receiveWindow"></param>
         /// <returns></returns>
         Task<DepositListResponse> GetDepositHistory(FundHistoryRequest request, int receiveWindow = 5000);
+
+        /// <summary>
+        /// Current exchange trading rules and symbol information
+        /// </summary>
+        Task<ExchangeInfoResponse> GetExchangeInfo();
+
     }
 }